enterprise: rework license summary caching
