+++ authentik/enterprise/api.py
@@ -1,6 +1,7 @@
 """Enterprise API Views"""
 
-from datetime import datetime, timedelta
+from dataclasses import asdict
+from datetime import timedelta
 
 from django.utils.timezone import now
 from django.utils.translation import gettext as _
@@ -8,7 +9,7 @@ from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import extend_schema, inline_serializer
 from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError
-from rest_framework.fields import BooleanField, CharField, DateTimeField, IntegerField
+from rest_framework.fields import CharField, IntegerField
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -19,18 +20,18 @@ from authentik.api.decorators import permission_required
 from authentik.core.api.used_by import UsedByMixin
 from authentik.core.api.utils import PassiveSerializer
 from authentik.core.models import User, UserTypes
-from authentik.enterprise.models import License, LicenseKey
+from authentik.enterprise.license import LicenseKey, LicenseSummarySerializer
+from authentik.enterprise.models import License
 from authentik.root.install_id import get_install_id
 
 
 class EnterpriseRequiredMixin:
     """Mixin to validate that a valid enterprise license
-    exists before allowing to safe the object"""
+    exists before allowing to save the object"""
 
     def validate(self, attrs: dict) -> dict:
         """Check that a valid license exists"""
-        total = LicenseKey.get_total()
-        if not total.is_valid():
+        if not LicenseKey.cached_summary().valid:
             raise ValidationError(_("Enterprise is required to create/update this object."))
         return super().validate(attrs)
 
@@ -59,19 +60,6 @@ class LicenseSerializer(ModelSerializer):
             "internal_users": {"read_only": True},
             "external_users": {"read_only": True},
         }
-
-
-class LicenseSummary(PassiveSerializer):
-    """Serializer for license status"""
-
-    internal_users = IntegerField(required=True)
-    external_users = IntegerField(required=True)
-    valid = BooleanField()
-    show_admin_warning = BooleanField()
-    show_user_warning = BooleanField()
-    read_only = BooleanField()
-    latest_valid = DateTimeField()
-    has_license = BooleanField()
 
 
 class LicenseForecastSerializer(PassiveSerializer):
@@ -111,31 +99,13 @@ class LicenseViewSet(UsedByMixin, ModelViewSet):
     @extend_schema(
         request=OpenApiTypes.NONE,
         responses={
-            200: LicenseSummary(),
+            200: LicenseSummarySerializer(),
         },
     )
     @action(detail=False, methods=["GET"], permission_classes=[IsAuthenticated])
     def summary(self, request: Request) -> Response:
         """Get the total license status"""
-        total = LicenseKey.get_total()
-        last_valid = LicenseKey.last_valid_date()
-        # TODO: move this to a different place?
-        show_admin_warning = last_valid < now() - timedelta(weeks=2)
-        show_user_warning = last_valid < now() - timedelta(weeks=4)
-        read_only = last_valid < now() - timedelta(weeks=6)
-        latest_valid = datetime.fromtimestamp(total.exp)
-        response = LicenseSummary(
-            data={
-                "internal_users": total.internal_users,
-                "external_users": total.external_users,
-                "valid": total.is_valid(),
-                "show_admin_warning": show_admin_warning,
-                "show_user_warning": show_user_warning,
-                "read_only": read_only,
-                "latest_valid": latest_valid,
-                "has_license": License.objects.all().count() > 0,
-            }
-        )
+        response = LicenseSummarySerializer(data=asdict(LicenseKey.cached_summary()))
         response.is_valid(raise_exception=True)
         return Response(response.data)
